% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychart.R
\name{psychart}
\alias{psychart}
\title{Plot psychart}
\usage{
psychart(
  x,
  labels = names(x),
  radius = 0.4,
  spiral.shift = 0,
  init.angle = 0,
  col = rainbow(length(x)),
  border = NA,
  lty = 1,
  lwd = 1,
  center = c(0, 0),
  tck.length = radius/10,
  norm.radius = TRUE,
  axes = FALSE,
  xlab = "",
  ylab = "",
  new = TRUE,
  N = 50,
  ...
)
}
\arguments{
\item{x}{a vector of non-negative numerical quantities. The values in x are displayed as the areas of pie slices.}

\item{labels}{character gibing names of slices}

\item{radius}{psypie radius in user coordinates (by x-axis, see norm.radius)}

\item{spiral.shift}{twisting size in degrees}

\item{init.angle}{number specifying the starting angle (in degrees) for the slices. Defaults to 0 (i.e., ‘3 o'clock’) unless clockwise is true where init.angle defaults to 90 (degrees), (i.e., ‘12 o'clock’).}

\item{col}{a vector of colors to be used in filling or shading the slices}

\item{border, lty, lwd}{paramteres to be passed to polygon}

\item{center}{center of psypie.}

\item{tck.length}{relative tick lenght (as fraction from radius)}

\item{norm.radius}{logical, should psypie appear as cycle in device coordinates (defoult). Psypie will be cycle in user coordinates if norm.radius is FALSE}

\item{axes}{logical whether axis shoul be plotted}

\item{xlab, ylab}{axis labels}

\item{new}{logical, should new plot be created (adds psypie to existing if FALSE)}

\item{N}{number of points used to plot each line segment. Set higher for better detalization.}

\item{...}{}
}
\description{
Creates new plot if center
}
